Fixed audio reading. Added fileutils.py for file moving. Rearranged tsv file writing for more correct and simplistic timing. Removed need fo labels on trial columns (added auto detecter).
