+++ BIDS_converter/utils/__init__.py
@@ -1,2 +1,1 @@
-from .utils import *
-from .organize import *
+from . import utils, organize, fileutils
