+++ BIDS_converter/utils/utils.py
@@ -2,16 +2,14 @@
 # -*- coding: utf-8 -*-
 
 import os
-import re
-import shutil
-import stat
+import subprocess
 from pathlib import Path
+from typing import Union, TypeVar
 
-import exrex as ex
 import numpy as np
 import pandas as pd
-from scipy.io import wavfile
-from typing import List, Union
+
+PathLike = TypeVar("PathLike", str, os.PathLike)
 
 
 class DisplayablePath:
@@ -68,7 +66,7 @@ class DisplayablePath:
     def _default_criteria(cls, path):
         return True
 
-    def displayable(self) -> Path | str:
+    def displayable(self) -> Union[Path, str]:
         if self.parent is None:
             return self.path
 
@@ -89,107 +87,12 @@ class DisplayablePath:
         return ''.join(reversed(parts))
 
 
-def match_regexp(config_regexp, filename, subtype=False):
-    delimiter_left = config_regexp["left"].replace("(", "(?:")
-    delimiter_right = config_regexp["right"].replace("(", "(?:")
-    match_found = False
-
-    if subtype:
-        for to_match in config_regexp["content"]:
-            if re.match(".*?"
-                        + delimiter_left
-                        + '(' + to_match[1].replace("(", "(?:") + ')'
-                        + delimiter_right
-                        + ".*?", filename):
-                match = to_match[0]
-                match_found = True
-    else:
-        for to_match in config_regexp["content"]:
-            if re.match(".*?"
-                        + delimiter_left
-                        + '(' + to_match.replace("(", "(?:") + ')'
-                        + delimiter_right
-                        + ".*?", filename):
-                match = re.match(".*?"
-                                 + delimiter_left
-                                 + '(' + to_match.replace("(", "(?:") + ')'
-                                 + delimiter_right
-                                 + ".*?", filename).group(1)
-                match_found = True
-    assert match_found
-    return match
-
-
-def gen_match_regexp(config_regexp, data,
-                     subtype=False):
-    """takes a match config and generates a matching string
-
-    :param config_regexp:
-    :type config_regexp:
-    :param data:
-    :type data:
-    :param subtype:
-    :type subtype:
-    :return:
-    :rtype:
-    """
-    if data.startswith("0"):
-        data = data.lstrip("0")
-    match_found = False
-    for to_match in config_regexp["content"]:
-        if re.match(to_match, data):
-            match_found = True
-    if not match_found:
-        raise AssertionError(
-            "{newname} doesn't match config criteria {given}".format(
-                newname=data, given=config_regexp["content"]))
-    left = ex.getone(config_regexp["left"])
-    right = ex.getone(config_regexp["right"])
-    newname = left + data + right
-
-    try:
-        if data == match_regexp(config_regexp, newname, subtype=subtype):
-            return newname
-        else:
-            raise ValueError("{newname} doesn't match config criteria".format(
-                newname=newname))
-    except AssertionError:
-        # return self.gen_match_regexp(config_regexp, data.lstrip("0"),subtype)
-        # except RecursionError:
-        raise AssertionError(
-            "{newname} doesn't match config criteria {given}".format(
-                newname=newname, given=config_regexp))
-
-
-def read_write_edf(read_obj, chn):
-    if isinstance(chn, str):
-        chn = read_obj.getSignalLabels().index(chn)
-    read_obj.readSignal(chn)
-
 
 # this part of the code creates the tree graphic
 def tree(path):
     paths = DisplayablePath.make_tree(Path(path))
     for path_to_display in paths:
         print(path_to_display.displayable())
-
-
-def rot_x(alpha):
-    return np.array([[1, 0, 0]
-                        , [0, np.cos(alpha), np.sin(alpha)]
-                        , [0, -np.sin(alpha), np.cos(alpha)]])
-
-
-def rot_y(alpha):
-    return np.array([[np.cos(alpha), 0, -np.sin(alpha)]
-                        , [0, 1, 0]
-                        , [np.sin(alpha), 0, np.cos(alpha)]])
-
-
-def rot_z(alpha):
-    return np.array([[np.cos(alpha), np.sin(alpha), 0]
-                        , [-np.sin(alpha), np.cos(alpha), 0]
-                        , [0, 0, 1]])
 
 
 def is_number(s: str) -> bool:
@@ -217,6 +120,13 @@ def is_number(s: str) -> bool:
         return False
 
 
+def bids_validator(bids_dir):
+    assert bids_dir is not None, "Cannot launch bids-validator wit" \
+                                       "hout specifying bids directory !"
+    subprocess.check_call(['bids-validator',
+                           bids_dir])
+
+
 def str2num(s: str) -> Union[float, str]:
     if is_number(s):
         return float(s)
@@ -230,141 +140,8 @@ def slice_time_calc(TR, sNum, totNum, delay):
     return tslice
 
 
-def delete_folder(pth: os.PathLike):
-    for sub in pth.iterdir():
-        if sub.is_dir():
-            delete_folder(sub)
-        else:
-            sub.unlink()
-    pth.rmdir()
-
-
 def set_default(obj):
     if isinstance(obj, set):
         return list(obj)
     raise TypeError
 
-
-def force_remove(mypath: os.PathLike):
-    x = 0
-    e = None
-    while os.path.isfile(mypath) or os.path.isdir(mypath):
-        x += 1
-        if os.path.isfile(mypath):
-            os.remove(mypath)
-        try:
-            if os.path.isdir(mypath):
-                delete_folder(Path(mypath))
-        except OSError:
-            try:
-                shutil.rmtree(mypath)
-            except PermissionError:
-                for root, dirs, files in os.walk(mypath, topdown=False):
-                    for file in files:
-                        fullfile = os.path.join(root, file)
-                        os.chmod(fullfile, stat.S_IWUSR)
-                        os.remove(fullfile)
-                    for dir in dirs:
-                        try:
-                            delete_folder(os.path.join(root, dir))
-                        except AttributeError:
-                            os.rmdir(os.path.join(root, dir))
-                shutil.rmtree(mypath, ignore_errors=True)
-            except Exception as e:
-                shutil.rmtree(mypath, ignore_errors=True)
-        if x >= 1000:
-            if e is not None:
-                raise RuntimeError(
-                    mypath + " could not remove all files or directories becau"
-                             "se of " + e)
-            else:
-                raise
-
-
-def eval_df(df: pd.DataFrame, exp: str,
-            file_dir: os.PathLike = "") -> pd.Series:
-    """input a df and expression and return a single dataframe column
-
-    :param df:
-    :type df:
-    :param exp:
-    :type exp:
-    :param file_dir:
-    :type file_dir:
-    :return:
-    :rtype:
-    """
-    if exp in df.columns:
-        return df[exp].squeeze()
-    for name in [i for i in re.split(r"[ +\-/*%]", exp) if i != '']:
-        if name in df.columns:
-            if is_number(df[name]):
-                df[name] = df[name].astype(float)
-            elif os.path.isfile(os.path.join(file_dir, str(df[name].iloc[0]))):
-                for i, (_, fname) in enumerate(df[name].iteritems()):
-                    fname = os.path.join(file_dir, fname)
-                    frames, data = wavfile.read(fname)
-                    duration = data.size / frames
-                    df[name].iat[i] = duration
-                df[name] = df[name].astype(float)
-            else:
-                df[name] = df[name]
-        elif not is_number(name):
-            if len([i for i in re.split(r"[ +\-/*%]", exp) if i != '']) > 1:
-                continue
-            return pd.Series([name] * df.shape[0], dtype="string")
-        else:
-            df[name] = pd.Series([float(name)] * df.shape[0], dtype="float")
-    return df.eval(exp).squeeze()
-
-
-def trigger_from_excel(filename: os.PathLike, participant: str) -> Union[int, str]:
-    """replace trigger channels with trigger label
-
-    :param filename:
-    :type filename:
-    :param participant:
-    :type participant:
-    :return:
-    :rtype:
-    """
-    xls_file = filename
-    xls_df = pd.ExcelFile(filename).parse(participant)
-
-    if any("Trigger" in column for column in xls_df):
-        for column in xls_df:
-            if "Trigger" in column:
-                trig_label = xls_df[column].iloc[0]
-                if is_number(trig_label):
-                    return int(trig_label)
-                else:
-                    return trig_label
-    else:
-        raise KeyError("'Trigger' not found in " + xls_file)
-
-
-def str2list(x: str) -> list:
-    ttable = "".maketrans({'[': '', ']': '', '\'': ''})
-    return [str2num(x) for x in x.translate(ttable).split()]
-
-
-def check_stims(stim_dir: os.PathLike, labels: pd.Series) -> pd.Series:
-    out_label = labels.copy()
-    files = os.listdir(stim_dir)
-    for i, label in enumerate(labels.tolist()):
-        if label is None:
-            out_label.iloc[i] = None
-        elif label in files:
-            out_label.iloc[i] = label
-        elif label + ".wav" in files:
-            out_label.iloc[i] = label + ".wav"
-        else:
-            out_label.iloc[i] = check_lower(label, files)
-    return out_label
-
-
-def check_lower(item: str, string_list: List[str]) -> str:
-    for stim_file in string_list:
-        if item in stim_file.lower():
-            return stim_file
-    raise FileNotFoundError("No stim files match {}".format(item))
